marketplace_backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
Pillow
django-cors-headers
